pyproject.toml
[tool.poetry]
name = "scrapdf"
version = "0.1.0"
description = "Python library for scraping PDF documents including text extraction, topic modelling, entity recognition, and table extraction."
authors = ["viralogic <bwfenske@ualberta.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"pdfminer.six" = "^20201018"
pytesseract = "^0.3.7"
pdf2image = "^1.14.0"
pillow = "^8.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-cov = "^2.11.1"
flake8 = "^3.8.4"
black = "^20.8b1"
pre-commit = "^2.9.3"
PyPDF2 = "^1.26.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
